extract wave selection route
